Explicitly grant START_ACTIVITIES_FROM_BACKGROUND to  com.android.phone.
It's needed because when we broadcast ACTION_SIM_SLOT_STATUS_CHANGED
we want to allow the receiving app to start an activity from the
background.

The app already has it implicitly, since it has the same shared UID as
com.android.stk which has the permission for unrelated reasons. Making
it explicit makes it less likely it will lose the permission
accidentally re-introducing a subtly bug.

@rc420head note: coral was not happy with this commit in frameworks/base, adding it here lets it boot again

Bug: 132691768
Test: Builds
Change-Id: I6a318b668b49fba1872b28a0dca75b7d57cb556e